% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{.recanting_twins_control}
\alias{.recanting_twins_control}
\title{Change Recanting Twins Estimator Hyperparameters}
\usage{
.recanting_twins_control(...)
}
\arguments{
\item{...}{}
}
\description{
Change Recanting Twins Estimator Hyperparameters
}
